Syllabus Page Automatically Jump to Current Week
This commit adds a new feature that when the syllabus page is clicked. The page
will automatically jump to most current week. This feature is helpful for
student because as the course progress, students need to scroll long blobs of
text and resources. The most common destination will be the current week. Thus
this new feature.

Note that this completely depends on the highlighting feature. It that breaks,
autojump will not work. If it does not work, it won't break the website.

- Simon Mo